add:OrderRequestProductBo 新增giveIntegral字段。赠送积分字段

+++ kxmall-system/src/main/java/com/kxmall/order/domain/bo/OrderRequestProductBo.java
@@ -22,7 +22,7 @@ public class OrderRequestProductBo {
 
     private Long integral;
 
-    private Long giveIntegral;
+    private BigDecimal giveIntegral;
